+++ packages/react-components/react-migration-v0-v9/library/src/components/Attachment/Attachment.test.tsx
@@ -1,5 +1,5 @@
 import '@testing-library/jest-dom';
-import { isConformant } from '@fluentui/react-conformance';
+// import { isConformant } from '@fluentui/react-conformance';
 import * as React from 'react';
 import { render, fireEvent } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
@@ -7,12 +7,13 @@ import userEvent from '@testing-library/user-event';
 import { Attachment } from './Attachment';
 
 describe('Attachment', () => {
-  isConformant({
-    Component: Attachment,
-    componentPath: module!.filename.replace('.test', ''),
-    displayName: 'Attachment',
-    disabledTests: ['has-docblock', 'has-top-level-file', 'component-has-static-classnames-object'],
-  });
+  // to prevent out of memory
+  // isConformant({
+  //   Component: Attachment,
+  //   componentPath: module!.filename.replace('.test', ''),
+  //   displayName: 'Attachment',
+  //   disabledTests: ['has-docblock', 'has-top-level-file', 'component-has-static-classnames-object'],
+  // });
 
   it('renders a default state', () => {
     const { getByText } = render(<Attachment>Actionable</Attachment>);