+++ packages/web-components/src/textarea/index.ts
@@ -1,0 +1,14 @@
+export { definition as TextAreaDefinition } from './textarea.definition.js';
+export { BaseTextArea, TextArea } from './textarea.js';
+export {
+  TextAreaAppearance,
+  TextAreaAppearancesForDisplayShadow,
+  TextAreaAutocomplete,
+  TextAreaResizableResize,
+  TextAreaHorizontallyResizableResize,
+  TextAreaVerticallyResizableResize,
+  TextAreaResize,
+  TextAreaSize,
+} from './textarea.options.js';
+export { styles as TextAreaStyles } from './textarea.styles.js';
+export { template as TextAreaTemplate } from './textarea.template.js';